# docs/core/specifications/api.openapi.yaml
openapi: 3.1.0
info:
  title: Chit API
  version: 0.2.0
  description: |
    Draft contract. Platform routes live under /api; each module owns /api/<module>/... (see docs/modules/<module>/contracts.md).
    Routes marked x-implemented exist in the hub today (core/server + module routes.py). No authentication is applied (ADR-0007, interim, loopback only).
servers:
  - url: /api
paths:
...
  /planner/reminders/{id}/skip:
    post: { operationId: skipReminderToday, x-implemented: true, x-module: planner/reminders, description: 'Body {skipped: boolean}; today only.', parameters: [{ name: id, in: path, required: true, schema: { type: string } }], responses: { '200': { description: Updated }, '404': { description: Not found } } }
...
  /planner/reminders:
    get: { operationId: listReminders, x-implemented: true, x-module: planner/reminders, responses: { '200': { description: Active reminders } } }
    post: { operationId: createReminder, x-implemented: true, x-module: planner/reminders, responses: { '201': { description: Created }, '400': { description: Invalid } } }
...
  /planner/reminders/{id}:
    put: { operationId: updateReminder, x-implemented: true, x-module: planner/reminders, parameters: [{ name: id, in: path, required: true, schema: { type: string } }], responses: { '200': { description: Saved }, '404': { description: Not found } } }
    delete: { operationId: archiveReminder, x-implemented: true, x-module: planner/reminders, parameters: [{ name: id, in: path, required: true, schema: { type: string } }], responses: { '200': { description: Archived }, '404': { description: Not found } } }
  /planner/reminders/{id}/done:
    post: { operationId: markReminderDone, x-implemented: true, x-module: planner/reminders, description: 'Body {done: boolean}. Mark a reminder done for today on the dashboard (or undo). The reminder stays listed; a child''s phone keeps its own marks.', parameters: [{ name: id, in: path, required: true, schema: { type: string } }], responses: { '200': { description: OK } } }
...
  /kids/grades/subjects/{id}:
    put: { operationId: updateSubject, x-implemented: true, x-module: kids/learning, description: 'Body {name?, code?, kind?: core|minor|elective}: full name, short code (1-6 characters, unique per child) and type. A new name is carried to the subject lessons, homework and bag items. A subject cannot be added or removed here: it exists because a lesson with its title is in the school-day plan (POST /kids/school/slots).', parameters: [{ name: id, in: path, required: true, schema: { type: string } }], responses: { '200': { description: OK } } }
    delete: { operationId: removeSubject, x-implemented: true, x-module: kids/learning, description: 'Remove a subject for good with its lessons in the plan, its grades and bag items; homework that named it loses the subject. Returns how many lessons and grades were removed.', parameters: [{ name: id, in: path, required: true, schema: { type: string } }], responses: { '200': { description: OK } } }
  /kids/grades/subjects/{id}/merge:
    post: { operationId: mergeSubject, x-implemented: true, x-module: kids/learning, description: 'Body {into: id}. Fold this duplicate subject into another of the same child: lessons, grades, homework and bag items move; the kept subject keeps its name, code and type.', parameters: [{ name: id, in: path, required: true, schema: { type: string } }], responses: { '200': { description: OK } } }
...
  /kids/grades/{id}:
    delete: { operationId: removeGrade, x-implemented: true, x-module: kids/learning, description: 'Delete a grade.', parameters: [{ name: id, in: path, required: true, schema: { type: string } }], responses: { '200': { description: OK } } }
...
  /kids/school/slots/{id}:
    put: { operationId: updateSchoolSlot, x-implemented: true, x-module: kids/school, description: 'Body {start, end, title, kind, note?, subject_kind?}. Renaming the last lesson of a subject renames the subject and keeps its grades.', parameters: [{ name: id, in: path, required: true, schema: { type: string } }], responses: { '200': { description: OK } } }
    delete: { operationId: removeSchoolSlot, x-implemented: true, x-module: kids/school, description: 'Remove a slot.', parameters: [{ name: id, in: path, required: true, schema: { type: string } }], responses: { '200': { description: OK } } }
...
  /kids/homework/{id}:
    put: { operationId: updateHomework, x-implemented: true, x-module: kids/learning, description: 'Body {done: boolean} ticks or reopens it; {dismissed: boolean} marks it not relevant or brings it back (done and dismissed exclude each other); a full body {kind, subject?, title, due_on, note?} edits it.', parameters: [{ name: id, in: path, required: true, schema: { type: string } }], responses: { '200': { description: OK } } }
    delete: { operationId: removeHomework, x-implemented: true, x-module: kids/learning, description: 'Remove a task.', parameters: [{ name: id, in: path, required: true, schema: { type: string } }], responses: { '200': { description: OK } } }
...
  /kids/bag/items/{id}:
    delete: { operationId: removeBagItem, x-implemented: true, x-module: kids/school, description: 'Remove a bag item.', parameters: [{ name: id, in: path, required: true, schema: { type: string } }], responses: { '200': { description: OK } } }
...